requirements.txt
asgiref==3.8.1
Django==4.2.11
django-js-asset==2.2.0
django-smart-selects==1.6.0
django-widget-tweaks==1.5.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
